// client/src/components/Navbar/Navbar.types.ts
export interface NavbarProps {
  show?: true | false;
}

export interface NavbarStyledProps {
  show: boolean;
  isSmall: boolean | null;
}

export interface UlStyledProps {
  isSmall: boolean | null;
  isOpen: boolean;
  show: boolean;
}

export interface NavMenuProps {
  isSmall: boolean | null;
}